test dir rename and add new sceanrio file with manual steps
