+++ apps/src/tests/single-feature-tests/stack-v5/test-stack-prevent-native-dismiss-nested-stack/index.tsx
